Move cucumber code around
